Usage: centralize stored token count sanitization
